Remove silent flags for cURLs and change to use Apache archive for .tar.gz download
Done for both 0.2.1 and 0.3.0 previous releases. I hope that using the
archive links isn't that big of a deal. It *seems* like
that is what they are there for, but I do not know for certain. I would
*really* like if NiFi had signed tags, and then I could just pull from Github
and build from source. I thought they were going to do this for 0.4.0 but it
doesn't look like that happened.