Create 0.4.0-openjdk-8 image using a mirror download URL and a user named 'nifi'
Going to use mirror for now. Also, created a user and group named 'nifi' so
that the service isn't running as root anymore. Will also create images for
other Java versions, and possibly just point them to the Apache archive URL.
Like mentioned before, it might be better if I communicate again with the
project and get them to create a signed tag, and then pulling from Github
instead of randomly choosing a mirror that is close to where I am.