Add .travis.yml with Docker installed and scripts for linting it
I added a Dockerfile that contains Ruby and a Travis linting gem. I wanted to
use the Alpine image, but I was getting Gem installation errors each time so I
decided to just not deal with it right now. I think the plan will be for each
build to generate, build, and tag a Docker image for a specific Nifi/Java version.
I can make use of the build matrix to fan it out, and some sort of templating
so that I can stop copying the same exact Dockerfile around.