Build out make_docker.py to write templates and add env variables to Travis build matrix
This next push will see if environment variables can be substituted in for
command line arguments. I'm guessing they can be, but we will see. There
is currently no building/pushing of the written dockerfile, that is the
next step.

Logging was also added to the script, and a logger is globally shared. The
level can be configured as a CL parameter.