Completely strip down Dockerfile template
Builds have been constantly failing on Docker Hub for this project.
The curl to the NiFi hosting of the zipped installation keeps exiting
with an error code '60' in different base images.

As part of these changes to scaffold out, I'm also going to take care of
https://github.com/mkobit/docker-nifi/issues/1 . I will pull from the Github tag,
check if the signature matches, and build it. The images might be heavy (still),
but https://github.com/mkobit/docker-nifi/issues/3 will lead be investigating using
Alpine for the images