Add a secure parameter to .travis.yml file for testing and other fixups
- Remove multiple "latest" tags.
- Added another build for Java9 and NiFi 0.5.1.
- TODOs around better organization and modularization (such as the docker tags)

The last build also proved that the export of these environment variables are
secure (stripped from logs) but any printing or logging of those variables is
in clear text. This commit will help with determining how .travis.yml secure
values are exported and also see if the subprocess command echoes out the
command that is executed.