Use a distribution_url as part of template properties
This is basically a copy of the existing Dockerfiles used for versions <= 0.5.
It has a few tweaks for the Python `string.Template` class, but is mainly the
same.

Building from source is a little bit more complicated than I hoped. Building
NiFI requires a Maven version >= 3.1. The Maven version that can be installed
on the java:8 base image is a 3.0 version, and the Maven enforcer plugin
makes sure of that by failing early.

I think when I do the investigation into an Alpine image (because these are
HUGE), I will do the "build-from-source" work, too.