Add the 0.5.0, 0.4.1, and 0.4.0 version to .travis.yml
Going to see how the build scales across Travis with this push. Also,
this will change most of the past images to use JRE over JDK. I was
unaware of that I wasn't only using the JRe before.

Another weird thing is the ImageLayers badges seems to be off, as some
of the images don't exactly exist.