Remove git_tag substitution paramatert and change up some of the distribution_urls
Some of the builds failed on Travis because those URLs don't host that
specific version. Needed to go through and change a few of them. Would
be great if I could at least cache the downloads, since a bunch of the
images are built using the same base layer of NiFi. I'll have to think
about this some more when it comes to the "building from source"
scenario.

Also, removed the git_tag parameter, as it was unused. The intent was
for the "build frm source" side, but I haven't figured out a great way
to do it when utilizing the Docker Hub _java_ base images.