Switch to include/exclude model
None of the versions < 0.5.x are being hosted on any of the mirrors.
This is making it hard to rebuild/publish older versions, especially
when newer versions of the Java images become available. This is another
point for "building from source".

I switched to the include/exclude model, so that all of those old images
are being skipped. This isn't ideal, but with NiFi still being a < 1.0.0
version, I don't think it is too bad that only the newer images are
being rebuilt at the time of this change.