Rewrite and add lots of documentation. Rename Travis utilities.
Bunch of additional documentation and clarification done. I also figured
out how to "embed" content into an Asciidoc list. They call it
["attaching"](http://asciidoctor.org/docs/asciidoc-writers-guide/#complex-list-content).
This is different to me than how Markdown does it. Markdown just uses
additional indentation, when Asciidoc uses the plus (+) sign to attach
the content to the list, which is actually more straightforward and
easier to understand.

[ci skip]