Move the group, user, and user ownership into a single command
This reduces the 'virtual size' by 480 MB, at least in the OpenJDK 8
JRE. I don't really understand what this is, but I _think_ it will
affect download times. The runtime size is not different, but the
layered commit size is much different. At least, I think that is what
will happen.