Clean up documentation - follow Asciidoc recommended practices
Following the ["One Sentence Per
Line](http://asciidoctor.org/docs/asciidoc-recommended-practices/#one-sentence-per-line), rather than following the 80 character arbitrary limit I was sort-of following before.
The reasons listed in that article are all entirely sound and are leading me to believe that it is a much better method than hard-capping lines.

[ci skip]